udah ngepull

+++ database/migrations/2023_10_04_065113_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('nama_kamar');
-            $table->decimal('harga');
+            $table->integer('harga');
             $table->text('deskripsi');
             $table->string('image');
             $table->timestamps();
